Hacker News Elixir/Erlang Notifier v0.0.1 HNE.HNStories

This module pulls a list of the n top news stories from the HN firebase directory of Top News Stories, it then checks each news story title for ‘Elixir’ or ‘Erlang’.

Functions

decode_response(arg)

Decode_response handles the httpoison get request from the firebase story link and parses the received JSON to a map.

Returns tuple { :ok, body } where the body is the article details in map format.

fetch()

Fetch retrieves list of top HN news stories.

firebase_url_get(story)

Firebase_url_get generates the firebase url for that particular story.

get_stories(arg1)

Get_stories retrieves each news article from the firebase locaiton and checks the title.

When the incoming story list is empty it waits ten minutes and then runs again.

handle_date()

builds a datetime string from :calendar.

handle_firebase_list_response(arg)

handle_firebase_list_response receives the HTTPoison response and formats it for further processing.

Returns a tuple { :ok, body } where body is a list of strings of article ID’s with some bits on the ends that needs to be trimmed.

handle_json(arg)

Handle_json gets the map from http response and checks if the title field in the map contains the words ‘erlang’ & ‘elixir’.

Returns tuple { :ok, body } where body is the map if it does contain the keywords and ‘nothing’ if the article is not “interesting”.

handle_map(arg)

handle_map handles the sending of the article or the printing of the result to the console.

If the article is elixir/erlang relevant the function prints the title of article and sends relevant email.

If article is not “interesting” it just sends “nothing” to console.

handle_string(arg)

Handle_string cleans up the firebase list (comes in as string).

Returns list of firebase story ID’s