Hacker News Elixir/Erlang Notifier v0.0.1 HNE.Mailer

Module sets up mailing config as well as function

Functions

config()

smtp config for sending email

deliver(email)

delivers mail.

sending_email(title, link)

sets up mail for delivery including new title and link.