Hacker News Elixir/Erlang Notifier v0.0.1 HNE.StoryId

This module keeps track of all story ids that have been checked in the past in a csv in the root folder.

Functions

find_val(list, id_new, res)

check if value ‘id_new’ exists in list

return_new(firebase_ids)

Return list of id’s that are not in the csv

return_new(current, list, new_ids)